sloelux-perfbot/requirements.txt
google-adk>=1.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
mcp-agent>=0.1.3
a2a>=0.44
requests>=2.32.0
python-dotenv>=1.1.0
pydantic>=2.11.0
aiohttp>=3.12.0
pillow        # for image conversion
psycopg2-binary  # for PostgreSQL
sentry-sdk      # for error tracking
datadog         # for metrics and monitoring 